+++ src/app/(dashboard)/instagram-feed/components/list/list.tsx
@@ -1,4 +1,4 @@
-import { getInstagramFeeds } from '@/lib/instagram-feed';
+import { getInstagramFeeds } from '@/lib/instagram-feed/get-feeds';
 
 import InstagramFeedListTable from './table';