+++ CHANGELOG.md
@@ -98,6 +98,7 @@ All notable changes to this project will be documented in this file.
 
 - split instagram feed actions file into create and update action files
 - feed form hook to use tanstack query mutation
+- feed list to fetch from supabase
 
 ###### Media